refactor: module project structure
