--- terminal-backend/src/main/java/org/fyp24064/repository/ChatMessageRepository.java
@@ -1,9 +1,0 @@
-package org.fyp24064.repository;
-
-import org.fyp24064.model.ChatMessage;
-//import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
-
-//public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long>  {
-//    List<ChatMessage> findAllByChatRoomId(Long id);
-//}
